# entitlement-service/entitlement-service-openapi.yml
openapi: 3.0.3
info:
  title: Entitlement Service API
  description: |
    NextGen Workflow Entitlement Service - Central authority for authentication, authorization, and user management.
    
    This service provides:
    - Session-based authentication
    - Cerbos policy-based authorization
    - User and role management
    - Business application management
    
    **Authentication**: All endpoints except `/api/auth/login` require session-based authentication via `X-Session-Id` header.
    
    **Authorization**: Fine-grained access control is provided through Cerbos policy engine integration.
  version: 1.0.0
  contact:
    name: NextGen Workflow Team
servers:
  - url: http://localhost:8081
    description: Development server
...
paths:
...
  /api/system-test/users/uuid-test:
    get:
      tags:
        - System Testing
      summary: Test UUID user operations
      description: Tests hybrid schema with UUID-based user operations and lookups
      responses:
        '200':
          description: UUID user test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalActiveUsers:
                    type: integer
                    example: 10
                  aliceFound:
                    type: boolean
                    example: true
                  bobFound:
                    type: boolean
                    example: true
                  aliceUserId:
                    type: string
                    format: uuid
                    example: "550e8400-e29b-41d4-a716-446655440001"
                  bobUserId:
                    type: string
                    format: uuid
                    example: "550e8400-e29b-41d4-a716-446655440002"
                  sampleUser:
                    $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
  /api/system-test/user-roles/uuid-test:
    get:
      tags:
        - System Testing
      summary: Test UUID user-role assignments
      description: Tests user-role assignment operations with UUID-based lookups
      responses:
        '200':
          description: UUID user-role test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  aliceUserId:
                    type: string
                    format: uuid
                    example: "550e8400-e29b-41d4-a716-446655440001"
                  totalRoles:
                    type: integer
                    example: 2
                  activeRoles:
                    type: integer
                    example: 2
                  roleDetails:
                    type: array
                    items:
                      type: object
                      properties:
                        roleName:
                          type: string
                          example: "INTAKE_ANALYST"
                        applicationName:
                          type: string
                          example: "onecms"
                        isActive:
                          type: boolean
                          example: true
        '500':
          description: Internal server error
  /api/system-test/authorization/engine-test:
    post:
      tags:
        - System Testing
      summary: Test authorization engine functionality
      description: Tests the hybrid authorization system with configurable parameters
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "alice.intake"
                  default: "alice.intake"
                resource:
                  type: string
                  example: "case"
                  default: "case"
                action:
                  type: string
                  example: "create"
                  default: "create"
      responses:
        '200':
          description: Authorization engine test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: "alice.intake"
                  userId:
                    type: string
                    format: uuid
                    example: "550e8400-e29b-41d4-a716-446655440001"
                  resource:
                    type: string
                    example: "case"
                  action:
                    type: string
                    example: "create"
                  authorizationResult:
                    type: object
                    properties:
                      allowed:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Access granted"
                      validationResult:
                        type: string
                        example: "none"
        '500':
          description: Internal server error
  /api/system-test/status:
    get:
      tags:
        - System Testing
      summary: Get system status and configuration
      description: Returns overall system health and entity counts for hybrid architecture validation
      responses:
        '200':
          description: System status information
          content:
            application/json:
              schema:
                type: object
                properties:
                  hybridSchema:
                    type: string
                    example: "active"
                  entityCounts:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                        example: 10
                      activeUsers:
                        type: integer
                        example: 10
                      totalRoleAssignments:
                        type: integer
                        example: 20
                      activeRoleAssignments:
                        type: integer
                        example: 18
                  systemReady:
                    type: boolean
                    example: true
                  timestamp:
                    type: integer
                    format: int64
                    example: 1640995200000
        '500':
          description: Internal server error
components:
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        username:
          type: string
          example: "alice.intake"
        email:
          type: string
          example: "alice.intake@company.com"
        firstName:
          type: string
          example: "Alice"
        lastName:
          type: string
          example: "Smith"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        attributes:
          type: object
          additionalProperties: true
          example:
            department: "HR"
            role: "INTAKE_ANALYST"
...